# api/openapi.json
{
 "openapi": "3.0.3",
 "info": {
 "title": "xrelease API",
 "description": "Self-hosted release notifier — management API and inbound webhooks.\n\nStarted with `xrelease serve`. Polling continues in the background; webhooks provide low-latency push for GitHub, GitLab, Gitea, Codeberg, Bitbucket, and Docker Hub.\n\nConfiguration is split: `bootstrap.toml` (infra) + `app/releases.yaml` (sources/teams/notifiers). Source kinds (`type` in app config): `github`, `codeberg`, `gitea`, `gitlab`, `bitbucket`, `docker`, `ghcr`, `quay`, `ecr`, `feed`, `pypi`, `npm`, `yarn`, `cargo`, `maven`, `nuget`, `hex`, `rubygems`, `packagist`, `cpan`, `artifacthub`.",
 "version": "0.1.0",
 "license": {
 "name": "Apache-2.0",
 "url": "https://www.apache.org/licenses/LICENSE-2.0"
 }
 },
 "servers": [
 {
 "url": "http://127.0.0.1:8080",
 "description": "Default local bind (`[api].listen`)"
 }
 ],
...
 "paths": {
...
 "/api/v1/status": {
 "get": {
 "tags": [
 "observability"
 ],
 "summary": "Process overview for dashboard UI",
 "operationId": "getStatus",
 "security": [
 {
 "bearerAuth": []
 }
 ],
 "responses": {
 "200": {
 "description": "Uptime, metrics snapshot, outbox depth",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/StatusResponse"
 }
 }
 }
 },
 "401": {
 "$ref": "#/components/responses/Unauthorized"
 }
 }
 }
 },
...
 "/api/v1/sources": {
 "get": {
 "tags": [
 "observability"
 ],
 "summary": "List configured sources with live runtime state",
 "operationId": "listSources",
 "security": [
 {
 "bearerAuth": []
 }
 ],
 "responses": {
 "200": {
 "description": "Config + poll state + per-source metrics",
 "content": {
 "application/json": {
 "schema": {
 "type": "array",
 "items": {
 "$ref": "#/components/schemas/SourceDetail"
 }
 }
 }
 }
 },
 "401": {
 "$ref": "#/components/responses/Unauthorized"
 }
 },
 "parameters": [
 {
 "name": "organization",
 "in": "query",
 "required": false,
 "schema": {
 "type": "string"
 },
 "description": "Limit results to one organization (namespaced ids/tags). Requires Viewer on that org."
 }
 ]
 }
 },
 "/api/v1/sources/{id}": {
 "get": {
 "tags": [
 "observability"
 ],
 "summary": "Get one source with runtime state",
 "operationId": "getSource",
 "security": [
 {
 "bearerAuth": []
 }
 ],
 "parameters": [
 {
 "name": "id",
 "in": "path",
 "required": true,
 "schema": {
 "type": "string",
 "example": "github:tokio-rs/tokio"
 }
 }
 ],
 "responses": {
 "200": {
 "description": "Source detail",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/SourceDetail"
 }
 }
 }
 },
 "401": {
 "$ref": "#/components/responses/Unauthorized"
 },
 "404": {
 "$ref": "#/components/responses/NotFound"
 }
 }
 }
 },
 "/api/v1/outbox": {
 "get": {
 "tags": [
 "observability"
 ],
 "summary": "Pending/failed notification outbox",
 "operationId": "listOutbox",
 "security": [
 {
 "bearerAuth": []
 }
 ],
 "parameters": [
 {
 "name": "limit",
 "in": "query",
 "required": false,
 "schema": {
 "type": "integer",
 "minimum": 1,
 "maximum": 200,
 "default": 50
 }
 },
 {
 "name": "organization",
 "in": "query",
 "required": false,
 "schema": {
 "type": "string"
 },
 "description": "Limit results to one organization (namespaced ids/tags). Requires Viewer on that org."
 }
 ],
 "responses": {
 "200": {
 "description": "Outbox entries (no message body)",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/OutboxListResponse"
 }
 }
 }
 },
 "401": {
 "$ref": "#/components/responses/Unauthorized"
 }
 }
 }
 },
 "/api/v1/outbox/requeue": {
 "post": {
 "tags": [
 "management"
 ],
 "summary": "Requeue dead-letter outbox notifications",
 "description": "Revive rows in `dead` status back to `pending` (same as CLI `xrelease outbox requeue`). Rate-limited with other mutating management routes.",
 "operationId": "requeueOutbox",
 "security": [
 {
 "bearerAuth": []
 }
 ],
 "responses": {
 "200": {
 "description": "Number of rows revived",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/OutboxRequeueResponse"
 }
 }
 }
 },
 "401": {
 "$ref": "#/components/responses/Unauthorized"
 },
 "403": {
 "description": "Authenticated but the principal's role is insufficient (RBAC).",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/ErrorResponse"
 }
 }
 }
 },
 "500": {
 "$ref": "#/components/responses/InternalError"
 }
 }
 }
 },
 "/api/v1/teams": {
 "get": {
 "tags": [
 "observability"
 ],
 "summary": "Team routing catalogue with per-team source counts",
 "operationId": "listTeams",
 "security": [
 {
 "bearerAuth": []
 }
 ],
 "responses": {
 "200": {
 "description": "Configured teams (`[[teams]]`) merged with source counts",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/TeamListResponse"
 }
 }
 }
 },
 "401": {
 "$ref": "#/components/responses/Unauthorized"
 }
 },
 "parameters": [
 {
 "name": "organization",
 "in": "query",
 "required": false,
 "schema": {
 "type": "string"
 },
 "description": "Limit results to one organization (namespaced ids/tags). Requires Viewer on that org."
 }
 ]
 }
 },
...
 },
 "components": {
 "securitySchemes": {
 "bearerAuth": {
 "type": "http",
 "scheme": "bearer",
 "description": "Static API key (`[api].api_key` / `XRELEASE_API_KEY`) or OIDC access token JWT (`[api.oidc]` / `XRELEASE_OIDC_ISSUER`). Required when either auth method is configured."
 }
 },
 "responses": {
 "Unauthorized": {
 "description": "Authentication failed",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/ErrorResponse"
 }
 }
 }
 },
...
 "NotFound": {
 "description": "No matching configured source",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/ErrorResponse"
 }
 }
 }
 },
 "InternalError": {
 "description": "Internal error",
 "content": {
 "application/json": {
 "schema": {
 "$ref": "#/components/schemas/ErrorResponse"
 }
 }
 }
 }
 },
 "schemas": {
...
 "ErrorResponse": {
 "type": "object",
 "properties": {
 "error": {
 "type": "string"
 }
 },
 "required": [
 "error"
 ]
 },
 "SourceSummary": {
 "type": "object",
 "properties": {
 "id": {
 "type": "string",
 "example": "github:tokio-rs/tokio"
 },
 "kind": {
 "type": "string",
 "example": "github",
 "description": "Provider kind: github, codeberg, gitea, gitlab, bitbucket, docker, ghcr, quay, ecr, feed, pypi, npm, yarn, cargo, maven, nuget, hex, rubygems, packagist, cpan, artifacthub"
 },
 "kind_label": {
 "type": "string"
 },
 "display_name": {
 "type": "string"
 },
 "interval_secs": {
 "type": "integer"
 },
 "jitter_secs": {
 "type": "integer"
 },
 "routing_tag": {
 "type": "string",
 "nullable": true,
 "description": "Team routing tag (matches notifier tags / [[teams]].tag)."
 },
 "notify_schedule": {
 "type": "string",
 "nullable": true,
 "description": "Cron expression (UTC) gating notification delivery; null = deliver immediately",
 "example": "0 9 * * MON-FRI"
 },
 "organization_id": {
 "type": "string",
 "nullable": true,
 "description": "Owning organization id when multi-org; null in legacy single-document mode."
 }
 },
 "required": [
 "id",
 "kind",
 "kind_label",
 "display_name",
 "interval_secs",
 "jitter_secs",
 "routing_tag",
 "notify_schedule"
 ]
 },
...
 "SeenReleaseView": {
 "type": "object",
 "properties": {
 "tag": {
 "type": "string"
 },
 "published_at": {
 "type": "string",
 "nullable": true
 },
 "url": {
 "type": "string",
 "nullable": true,
 "description": "Canonical upstream release page (stored at sync time or derived from provider template)"
 },
 "first_seen_at": {
 "type": "string"
 }
 },
 "required": [
 "tag",
 "first_seen_at"
 ]
 },
 "SourceDetail": {
 "allOf": [
 {
 "$ref": "#/components/schemas/SourceSummary"
 },
 {
 "type": "object",
 "properties": {
 "initialized": {
 "type": "boolean"
 },
 "last_polled_at": {
 "type": "string",
 "nullable": true
 },
 "seen_count": {
 "type": "integer",
 "format": "int64"
 },
 "latest_release_tag": {
 "type": "string",
 "nullable": true
 },
 "source_page_url": {
 "type": "string",
 "nullable": true,
 "description": "Upstream catalog page (repo, package index, feed URL)"
 },
 "seen_releases": {
 "type": "array",
 "items": {
 "$ref": "#/components/schemas/SeenReleaseView"
 }
 },
 "polls": {
 "type": "integer",
 "format": "int64"
 },
 "polls_not_modified": {
 "type": "integer",
 "format": "int64"
 },
 "poll_errors": {
 "type": "integer",
 "format": "int64"
 },
 "notifications": {
 "type": "integer",
 "format": "int64"
 }
 },
 "required": [
 "initialized",
 "last_polled_at",
 "seen_count",
 "latest_release_tag",
 "seen_releases",
 "polls",
 "polls_not_modified",
 "poll_errors",
 "notifications"
 ]
 }
 ]
 },
 "MetricsSnapshot": {
 "type": "object",
 "properties": {
 "polls_total": {
 "type": "integer",
 "format": "int64"
 },
 "polls_not_modified": {
 "type": "integer",
 "format": "int64"
 },
 "poll_errors": {
 "type": "integer",
 "format": "int64"
 },
 "notifications_total": {
 "type": "integer",
 "format": "int64"
 },
 "webhooks_accepted": {
 "type": "integer",
 "format": "int64"
 },
 "webhooks_ignored": {
 "type": "integer",
 "format": "int64"
 },
 "webhooks_duplicates": {
 "type": "integer",
 "format": "int64"
 },
 "webhooks_errors": {
 "type": "integer",
 "format": "int64"
 },
 "config_apply_total": {
 "type": "integer",
 "format": "int64"
 },
 "config_apply_rejected_total": {
 "type": "integer",
 "format": "int64"
 },
 "notify_breaker_skips": {
 "type": "integer",
 "format": "int64"
 },
 "outbox_enqueued_total": {
 "type": "integer",
 "format": "int64"
 },
 "outbox_delivery_failures_total": {
 "type": "integer",
 "format": "int64"
 },
 "outbox_dead_lettered_total": {
 "type": "integer",
 "format": "int64"
 },
 "outbox_requeued_total": {
 "type": "integer",
 "format": "int64"
 },
 "http_rate_limited_total": {
 "type": "integer",
 "format": "int64"
 },
 "prune_deleted_total": {
 "type": "integer",
 "format": "int64"
 }
 },
 "required": [
 "polls_total",
 "polls_not_modified",
 "poll_errors",
 "notifications_total",
 "webhooks_accepted",
 "webhooks_ignored",
 "webhooks_duplicates",
 "webhooks_errors",
 "config_apply_total",
 "config_apply_rejected_total",
 "notify_breaker_skips",
 "outbox_enqueued_total",
 "outbox_delivery_failures_total",
 "outbox_dead_lettered_total",
 "outbox_requeued_total",
 "http_rate_limited_total",
 "prune_deleted_total"
 ]
 },
 "StatusResponse": {
 "type": "object",
 "properties": {
 "version": {
 "type": "string"
 },
 "uptime_secs": {
 "type": "integer",
 "format": "int64"
 },
 "database_ok": {
 "type": "boolean"
 },
 "sources_configured": {
 "type": "integer"
 },
 "outbox_pending": {
 "type": "integer"
 },
 "outbox_dead": {
 "type": "integer"
 },
 "metrics": {
 "$ref": "#/components/schemas/MetricsSnapshot"
 },
 "config_apply": {
 "$ref": "#/components/schemas/ConfigApplyStatus"
 },
 "outbox_deferred": {
 "type": "integer",
 "description": "Pending/failed rows held until deliver_after (notify_schedule)."
 },
 "breakers": {
 "type": "array",
 "items": {
 "$ref": "#/components/schemas/BreakerOpenView"
 }
 }
 },
 "required": [
 "version",
 "uptime_secs",
 "database_ok",
 "sources_configured",
 "outbox_pending",
 "outbox_dead",
 "metrics",
 "config_apply",
 "outbox_deferred",
 "breakers"
 ]
 },
 "ConfigApplyStatus": {
 "type": "object",
 "properties": {
 "api_config_enabled": {
 "type": "boolean",
 "description": "Whether POST apply/rollback routes are mounted ([config_api].api_config)."
 },
 "ui_config_enabled": {
 "type": "boolean",
 "description": "Whether the dashboard form editor is on (ui_config + api_config + source=api)."
 },
 "config_source": {
 "type": "string",
 "enum": [
 "local",
 "api"
 ],
 "description": "Declared desired-state authority ([config_api].source)."
 },
 "hmac_configured": {
 "type": "boolean"
 },
 "desired_source": {
 "type": "string",
 "enum": [
 "ledger",
 "app_file",
 "empty"
 ]
 },
 "bootstrap_path": {
 "type": "string"
 },
 "revision": {
 "type": "integer",
 "format": "int64",
 "nullable": true
 },
 "content_sha256": {
 "type": "string",
 "nullable": true
 },
 "revision_label": {
 "type": "string",
 "nullable": true
 },
 "applied_at": {
 "type": "string",
 "format": "date-time",
 "nullable": true
 },
 "last_rejected_revision": {
 "type": "integer",
 "format": "int64",
 "nullable": true
 },
 "last_rejected_error": {
 "type": "string",
 "nullable": true
 }
 },
 "required": [
 "api_config_enabled",
 "ui_config_enabled",
 "config_source",
 "hmac_configured",
 "desired_source",
 "bootstrap_path"
 ]
 },
...
 "OutboxEntry": {
 "type": "object",
 "properties": {
 "id": {
 "type": "integer",
 "format": "int64"
 },
 "source_id": {
 "type": "string"
 },
 "identity": {
 "type": "string"
 },
 "status": {
 "type": "string",
 "enum": [
 "pending",
 "failed",
 "dead"
 ]
 },
 "attempts": {
 "type": "integer"
 },
 "last_error": {
 "type": "string",
 "nullable": true
 },
 "created_at": {
 "type": "string"
 },
 "title": {
 "type": "string"
 },
 "url": {
 "type": "string",
 "nullable": true
 },
 "routing_tag": {
 "type": "string",
 "nullable": true,
 "description": "Team routing tag carried on the event."
 },
 "deliver_after": {
 "type": "string",
 "format": "date-time",
 "nullable": true,
 "description": "Delivery is held until this moment (notify_schedule); null = deliver as soon as possible"
 },
 "organization_id": {
 "type": "string",
 "nullable": true,
 "description": "Derived from namespaced source_id (org::…)."
 }
 },
 "required": [
 "id",
 "source_id",
 "identity",
 "status",
 "attempts",
 "last_error",
 "created_at",
 "title",
 "url",
 "routing_tag",
 "deliver_after"
 ]
 },
 "OutboxListResponse": {
 "type": "object",
 "properties": {
 "entries": {
 "type": "array",
 "items": {
 "$ref": "#/components/schemas/OutboxEntry"
 }
 }
 },
 "required": [
 "entries"
 ]
 },
 "OutboxRequeueResponse": {
 "type": "object",
 "properties": {
 "requeued": {
 "type": "integer",
 "minimum": 0,
 "description": "Number of dead outbox rows reset to pending"
 }
 },
 "required": [
 "requeued"
 ]
 },
 "TeamView": {
 "type": "object",
 "required": [
 "tag",
 "source_count"
 ],
 "properties": {
 "tag": {
 "type": "string",
 "description": "Routing tag shared by sources (`routing_tag`) and notifiers (`tags`)"
 },
 "name": {
 "type": "string",
 "nullable": true,
 "description": "Human-readable label from [[teams]].name"
 },
 "source_count": {
 "type": "integer",
 "description": "Number of configured sources routing to this tag"
 }
 }
 },
 "TeamListResponse": {
 "type": "object",
 "properties": {
 "teams": {
 "type": "array",
 "items": {
 "$ref": "#/components/schemas/TeamView"
 }
 }
 },
 "required": [
 "teams"
 ]
 },
...
 "BreakerOpenView": {
 "type": "object",
 "properties": {
 "kind": {
 "type": "string",
 "description": "Backend kind (apprise, webhook, …)"
 },
 "label": {
 "type": "string",
 "description": "Operator-facing sink label"
 },
 "open": {
 "type": "boolean",
 "description": "True when the breaker is rejecting calls"
 }
 },
 "required": [
 "kind",
 "label",
 "open"
 ]
 }
 }
 }
}